+++ CHANGES.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## v0.16.0
+
+Additions:
+* Added lots of new back-end API functionality to the `vmwarelib.tagging.VmwareTagging` interface.
+* Added new `vsphere.tag_attach_bulk` action to attach a tag to a bulk set of resources.
+  Primary usecase is tagging all VMs in a cluster dynamically.
+* Added new `vsphere.tag_detach_bulk` action to detach a tag from a bulk set of resources.
+  Primary usecase is remoave a tag from all VMs in a cluster dynamically.
+
+Contributed by Nick Maludy (@nmaludy Encore Technologies).
+
 ## v0.15.1
 
 Changed action: